These codes are for lamallar fragments.
Explicit 4th order code. If mex is to be used then download the entire files, use matlab coder and generate mex files for the functions which are iterated in the main.m 's "for" loop. 
At present, these are working fine with out any issues, but needs the implementation of a repulsion potential and then the rescaling technique.